Improve coverage (part 1) (#44)
* Add more tests

* Add tests for Vault / proxy

* Add tests for MarketBorrowRateFeed

+++ src/oracle/IMarketBorrowRateFeed.sol
@@ -3,6 +3,7 @@ pragma solidity 0.8.24;
 
 /// @title IMarketBorrowRateFeed
 interface IMarketBorrowRateFeed {
-    /// @notice Returns the market borrow rate
-    function getMarketBorrowRate() external view returns (uint256);
+    /// @notice Returns the market borrow rate with 18 decimals
+    /// @dev Ex, returns 0.05e18 for 5%
+    function getMarketBorrowRate() external view returns (uint128);
 }
